tsrcmgr/requirements.txt
alabaster==0.7.12
arrow==1.2.2
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
binaryornot==0.4.4
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==7.1.2
commonmark==0.9.1
cookiecutter==2.1.1
coverage==6.4.1
cryptography==37.0.4
decorator==5.1.1
deepdiff==5.8.1
dictdiffer==0.9.0
distlib==0.3.4
docutils==0.17.1
executing==0.8.3
filelock==3.7.1
flake8==4.0.1
flake8-docstrings==1.6.0
idna==3.3
imagesize==1.4.1
importlib-metadata==4.12.0
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
jeepney==0.8.0
Jinja2==3.1.2
jinja2-time==0.2.0
keyring==23.6.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
ordered-set==4.1.0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
pip-check-reqs==2.3.2
pip-licenses==3.5.4
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.30
PTable==0.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.12.0
pyparsing==3.0.9
pytest==6.2.5
pytest-cov==3.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-pythonpath==0.7.4
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
PyYAML==6.0
readme-renderer==35.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.4
SecretStorage==3.3.2
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.0.2
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.3.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tox==3.25.1
traitlets==5.3.0
# Editable Git install with no remote (tsrcmgr==0.0.1)
-e /home/av/repos/tsrcmgr/tsrcmgr
twine==4.0.1
urllib3==1.26.10
virtualenv==20.15.1
wcwidth==0.2.5
webencodings==0.5.1
yamllint==1.26.3
zipp==3.8.0
